# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API em Go para encriptar/decriptar dados com Postgres",
        "title": "API Encode/Decode Example",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/decode/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Decode data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Token"
                        }
                    }
                }
            }
        },
        "/encode": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Encode data",
                "parameters": [
                    {
                        "description": "Dados",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.EncodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.EncodeResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.EncodeRequest": {
            "type": "object",
            "required": [
                "batch",
                "date",
                "time"
            ],
            "properties": {
                "batch": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "main.EncodeResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "main.Token": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        }
    }
}